% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{as.data.frame,ScoreResult-method}
\alias{as.data.frame,ScoreResult-method}
\title{as.data.frame method}
\usage{
\S4method{as.data.frame}{ScoreResult}(x)
}
\arguments{
\item{x}{a object of class ScoreResult}
}
\description{
as.data.frame method
}
\author{
Yuanlong Hu
}
