% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,ScoreResult-method}
\title{Coerce a ScoreResult object into a data frame}
\usage{
\S4method{as.data.frame}{ScoreResult}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A ScoreResult object}

\item{row.names}{NULL or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{logical. If TRUE, setting row names and converting column names (to syntactic names: see make.names) is optional. Note that all of R's base package as.data.frame() methods use optional only for column names treatment, basically with the meaning of data.frame(*, check.names = !optional). See also the make.names argument of the matrix method.}

\item{...}{other arguments}
}
\description{
Coerce a ScoreResult object into a data frame
}
\author{
Yuanlong Hu
}
