% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{BasicData-class}
\alias{BasicData-class}
\title{Class \code{BasicData}
This class represents the basic input data.}
\description{
Class \code{BasicData}
This class represents the basic input data.
}
\section{Slots}{

\describe{
\item{\code{drugnet}}{A directed graph}

\item{\code{vertices}}{Vertices of drug graph.}

\item{\code{diseasenet}}{Disease network.}

\item{\code{biomarker}}{Disease-related gene.}
}}

\author{
Yuanlong Hu
}
