% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{BioDescr-class}
\alias{BioDescr-class}
\title{Class \code{BioDescr}
This class represents the biological descriptor data.}
\description{
Class \code{BioDescr}
This class represents the biological descriptor data.
}
\section{Slots}{

\describe{
\item{\code{drug_geneset}}{from drug to geneset.}

\item{\code{geneset_gene}}{from geneset to gene for each drug.}

\item{\code{anno}}{Geneset ID and description.}
}}

\author{
Yuanlong Hu
}
