% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBasicData.R
\name{CreateBasicData}
\alias{CreateBasicData}
\title{CreateBasicData}
\usage{
CreateBasicData(..., diseasenet = NULL, biomarker = NULL)
}
\arguments{
\item{...}{Drug graph from \code{PrepareData}.}

\item{diseasenet}{A graph of Disease-related gene from \code{PrepareData}.}

\item{biomarker}{Character vector, the vector of Disease-related gene.}
}
\value{
A BasicData object.
}
\description{
Create BasicData Object
}
\author{
Yuanlong Hu
}
