% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPanalysis.R
\name{createNP}
\alias{createNP}
\title{createNP}
\usage{
createNP(data, DisNet)
}
\arguments{
\item{data}{BasicData object.}

\item{DisNet}{data frame}
}
\value{
a data frame
}
\description{
NP analysis.
}
\author{
Yuanlong Hu
}
