% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_network.R
\name{diff_network_char}
\alias{diff_network_char}
\title{diff_network_char}
\usage{
diff_network_char(graph1, graph2, output_all = FALSE)
}
\arguments{
\item{graph1}{A igraph object.}

\item{graph2}{A igraph object.}

\item{output_all}{FALSE}
}
\value{
A number vector.
}
\description{
Calculate the difference of network characters in two network
}
\author{
Yuanlong Hu
}
