% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugEA.R
\name{DrugEA}
\alias{DrugEA}
\title{DrugEA}
\usage{
DrugEA(
  gene,
  internal_geneset = "L1000_UP",
  input_geneset,
  output_df = TRUE,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  qvalueCutoff = 0.2
)
}
\arguments{
\item{gene}{a vector of gene symbol}

\item{internal_geneset}{one of "ImmGenTop150", "L1000_DN", "L1000_UP", "CMAP_DN", "CMAP_UP"}

\item{input_geneset}{input_geneset is a data frame or list, when internal_geneset is NULL.}

\item{output_df}{whether to return to the data frame}

\item{pvalueCutoff}{pvalue cutoff on enrichment tests to report}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{qvalueCutoff}{qvalue cutoff on enrichment tests to report as significant. Tests must pass i) pvalueCutoff on unadjusted pvalues, ii) pvalueCutoff on adjusted pvalues and iii) qvalueCutoff on qvalues to be reported.}
}
\value{
a data frame or enrichResult object
}
\description{
Enrichment analysis for drug target
}
\author{
Yuanlong Hu
}
