% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immcp-package.R
\docType{data}
\name{drugSample}
\alias{drugSample}
\title{A demo dataset contains a sample of herbal prescription.}
\description{
A demo dataset contains a sample of herbal prescription.
}
\keyword{datasets}
