% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDisDrugNet.R
\name{exportCytoscape}
\alias{exportCytoscape}
\title{Export an xlsx file to Cytoscape}
\usage{
exportCytoscape(graph, file)
}
\arguments{
\item{graph}{igraph object.}

\item{file}{file}
}
\value{
A workbook object
}
\description{
Export an xlsx file to Cytoscape.
}
\author{
Yuanlong Hu
}
