% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DEG.R
\name{get_DEG}
\alias{get_DEG}
\title{get_DEG}
\usage{
get_DEG(data, pdata, contrasts)
}
\arguments{
\item{data}{a data frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{pdata}{a character vector of phenotype}

\item{contrasts}{a character vector specifying contrasts}
}
\value{
list object
}
\description{
Select Top gene as biomarker
}
\author{
Yuanlong Hu
}
