% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list.R
\name{get_list}
\alias{get_list}
\title{get_list}
\usage{
get_list(dataframe)
}
\arguments{
\item{dataframe}{A data frame of drug target. The first column is drug name, and the second column is drug target.}
}
\value{
list
}
\description{
Create a new list from a data.frame of drug target and disease biomarker as input
}
\author{
Yuanlong Hu
}
