% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result.R
\name{get_result}
\alias{get_result}
\title{get_result}
\usage{
get_result(result, pvalueCutoff = 0.05)
}
\arguments{
\item{result}{an object of class ScoreResult.}

\item{pvalueCutoff}{p-value cutoff.}
}
\value{
a data.frame
}
\description{
Extract a table of the score result
}
\examples{
\dontrun{
  data("drugSample")
  FP <- extrFP(disease_biomarker = drugSample$disease_biomarker,
               drug_target = drugSample$herb_target,
               geneset = "ImmGenTop150")
  res <- score_fp(FP, n=100)
  res <- get_result(res)
}
}
\author{
Yuanlong Hu
}
