% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrFP.R
\name{getF}
\alias{getF}
\title{getF}
\usage{
getF(expr, pdata, level = "gene", withTentative = TRUE, geneset)
}
\arguments{
\item{expr}{A matrix of expression values where rows correspond to genes and columns correspond to samples.}

\item{pdata}{A character of phenotype.}

\item{level}{one of the gene or pathway}

\item{withTentative}{If set to TRUE, Tentative attributes will be also returned.}

\item{geneset}{A data frame of geneset containing two columns.}
}
\value{
A character of features.
}
\description{
Select features related to phenotype using Boruta
}
\author{
Yuanlong Hu
}
