% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{head,ScoreResult-method}
\alias{head,ScoreResult-method}
\title{head method}
\usage{
\S4method{head}{ScoreResult}(x, ...)
}
\arguments{
\item{x}{a object of class ScoreResult}

\item{...}{other arguments}
}
\description{
head method
}
\author{
Yuanlong Hu
}
