% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{head}
\alias{head}
\alias{head,ScoreResult-method}
\title{Return the First Parts of a ScoreResult Object}
\usage{
\S4method{head}{ScoreResult}(x, ...)
}
\arguments{
\item{x}{A ScoreResult object}

\item{...}{other arguments}
}
\description{
Return the First Parts of a ScoreResult Object
}
