% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFP.R
\name{Heatmap_simFP}
\alias{Heatmap_simFP}
\title{Heatmap_simFP}
\usage{
Heatmap_simFP(data, ...)
}
\arguments{
\item{data}{A drug fingerprints similarity matrix}
}
\value{
a ggplot2
}
\description{
Visualization of a drug fingerprints similarity matrix using ggplot2
}
\author{
Yuanlong Hu
}
