% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/score_network.R
\name{imm_centr}
\alias{imm_centr}
\alias{imm_centr,data.frame-method}
\alias{imm_centr,ScoreResultNet-method}
\alias{imm_centr.data.frame}
\alias{imm_centr.ScoreResultNet}
\title{imm_centr}
\usage{
imm_centr(x, ...)

\S4method{imm_centr}{data.frame}(x)

\S4method{imm_centr}{ScoreResultNet}(x, drug, node = "target", net = "disease")

imm_centr.data.frame(x, node, net)

imm_centr.ScoreResultNet(x, drug, node, net)
}
\arguments{
\item{x}{ScoreResultNet or data.frame object.}

\item{...}{additional parameters}

\item{drug}{drug name}

\item{node}{Nodes that need to be evaluated. one of "disease" and "target.}

\item{net}{Network. one of "disease" and "target.}
}
\value{
data.frame or ScoreResultNet object
}
\description{
Computing the centrality of complex networks
}
\author{
Yuanlong Hu

Yuanlong Hu

Yuanlong Hu
}
