% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immcp-package.R
\docType{package}
\name{immcp-package}
\alias{immcp-package}
\title{Candidate Prescriptions Discovery Based on Pathway Fingerprint}
\description{
The pathway fingerprint is a method to indicate the profile of significant pathways
being influenced by drugs, which may hint drug functions. Through the similarity of
pathway fingerprints, the potential relationship between disease and prescription
can be found. Ye (2012) <doi: 10.1007/s13238-012-2011-z>.
}
