% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_char.R
\name{natural_connectivity}
\alias{natural_connectivity}
\title{natural_connectivity}
\usage{
natural_connectivity(graph)
}
\arguments{
\item{graph}{A igraph object.}
}
\value{
A numeric vector.
}
\description{
Calculate the natural connectivity
}
\author{
Yuanlong Hu
}
