% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_network.R
\name{network_char_change}
\alias{network_char_change}
\title{network_char_change}
\usage{
network_char_change(DNet, target, output_all = TRUE)
}
\arguments{
\item{DNet}{A data frame of disease network containing two columns.}

\item{target}{character; drug target.}

\item{output_all}{output all result}
}
\value{
a data frame
}
\description{
Calculate the change of network characters
}
\examples{

  data("drugSample")
  network_char_change(DNet=drugSample$disease_network,
                      target=unique(drugSample$herb_target[,2]))
}
\author{
Yuanlong Hu
}
