% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_characterstics.R
\name{network_node_ks}
\alias{network_node_ks}
\title{network_node_ks}
\usage{
network_node_ks(graph, target = NULL, replicate = 1000)
}
\arguments{
\item{graph}{list; network node characters}

\item{target}{drug target}

\item{replicate}{the number of conduct bootstrapping sampling replications}
}
\value{
a data frame
}
\description{
Kolmogorov-Smirnov tests for node characters between networks
}
\examples{

data("drugSample")
graph <- igraph::graph.data.frame(drugSample$disease_network)
network_node_ks(graph, target=drugSample$disease_network[1:20,1])
}
\author{
Yuanlong Hu
}
