% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{overlap_pathway}
\alias{overlap_pathway}
\title{overlap_pathway}
\usage{
overlap_pathway(FP, Drug)
}
\arguments{
\item{FP}{A ScoreFP Object}

\item{Drug}{The drug names}
}
\value{
a vector or data frame
}
\description{
Performs set intersection on pathyways fingerprints
}
\author{
Yuanlong Hu
}
