% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDisDrugNet.R
\name{plot_BasicData}
\alias{plot_BasicData}
\title{plot_BasicData}
\usage{
plot_BasicData(
  BasicData,
  drug = NULL,
  disease = NULL,
  color = c(drug = "blue", herb = "lightblue", target = "orange"),
  ...
)
}
\arguments{
\item{BasicData}{BasicData object.}

\item{drug}{drug}

\item{disease}{disease}

\item{color}{Color}

\item{...}{Arguments}
}
\value{
Returns NULL, invisibly.
}
\description{
Plot Disease-Drug Network
}
\author{
Yuanlong Hu
}
