% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BioDescr.R
\name{plot_BioDescr}
\alias{plot_BioDescr}
\title{Plot Biological descriptor}
\usage{
plot_BioDescr(
  BioDescr,
  type = "heatmap",
  cluster_k = 2,
  colors = c("#2E9FDF", "#E7B800")
)
}
\arguments{
\item{BioDescr}{BioDescr object.}

\item{type}{one of "heatmap" and "clusterplot".}

\item{cluster_k}{Number vector, number of cluster.}

\item{colors}{vector of colors.}
}
\value{
Returns NULL, invisibly.
}
\description{
Plot Biological descriptor
}
