% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/plot_graph.R
\name{plot_graph}
\alias{plot_graph}
\alias{plot_graph,BasicData-method}
\alias{plot_graph,igraph-method}
\alias{plot_graph,HerbResult-method}
\title{Plot Disease-Drug Network}
\usage{
plot_graph(
  graph,
  drug,
  disease,
  Isolated = TRUE,
  vis = "visNetwork",
  color = c(drug = "#cca4e3", herb = "#ff461f", compound = "#ffc773", target =
    "#70f3ff"),
  width = 1,
  size = 20,
  ...
)

\S4method{plot_graph}{BasicData}(
  graph,
  drug,
  disease,
  Isolated = TRUE,
  vis = "visNetwork",
  color = c(drug = "#cca4e3", herb = "#ff461f", compound = "#ffc773", target =
    "#70f3ff"),
  width = 1,
  size = 20,
  ...
)

\S4method{plot_graph}{igraph}(
  graph,
  drug,
  disease,
  Isolated = TRUE,
  vis = "visNetwork",
  color = c(drug = "#cca4e3", herb = "#ff461f", compound = "#ffc773", target =
    "#70f3ff"),
  width = 1,
  size = 20,
  ...
)

\S4method{plot_graph}{HerbResult}(
  graph,
  drug,
  disease,
  Isolated = TRUE,
  vis = "visNetwork",
  color = c(drug = "#cca4e3", herb = "#ff461f", compound = "#ffc773", target =
    "#70f3ff"),
  width = 1,
  size = 20,
  ...
)
}
\arguments{
\item{graph}{graph.}

\item{drug}{drug.}

\item{disease}{disease.}

\item{Isolated}{Whether to delect Isolated nodes.}

\item{vis}{one of "igraph", "visNetwork" and "shiny".}

\item{color}{Nodes Color}

\item{width}{Edges width}

\item{size}{Nodes size}

\item{...}{Arguments}
}
\value{
Returns NULL, invisibly.
}
\description{
Plot Disease-Drug Network
}
\author{
Yuanlong Hu
}
