% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_KEGG_network}
\alias{plot_KEGG_network}
\title{plot_KEGG_network}
\usage{
plot_KEGG_network(FP, Drug, layout = "layout_in_circle")
}
\arguments{
\item{FP}{A ScoreFP object}

\item{Drug}{A character of drug name}

\item{layout}{use layout of igraph}
}
\value{
visNetwork
}
\description{
Drug-KEGG Pathyway network visualization
}
\author{
Yuanlong Hu
}
