% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_geneset.R
\name{read_gmt}
\alias{read_gmt}
\title{write_gmt}
\usage{
read_gmt(gmtfile, out_dataframe = TRUE)
}
\arguments{
\item{gmtfile}{A GMT file name or URL containing gene sets.}

\item{out_dataframe}{TRUE or FALSE}
}
\value{
data.frame, list
}
\description{
parse gmt file to a data.frame
}
\author{
Yuanlong Hu
}
