% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_rank.R
\name{res_rank}
\alias{res_rank}
\title{res_rank}
\usage{
res_rank(..., method = "RRA")
}
\arguments{
\item{...}{ScoreResult Object}

\item{method}{rank aggregation method, by defaylt \code{'RRA'}, other options are
\code{'min'}, \code{'geom.mean'}, \code{'mean'}, \code{'median'} and \code{'stuart'}}
}
\value{
a dataframe with two column
}
\description{
Rank the results by rank aggregation methods
}
\references{
Kolde, R., Laur, S., Adler, P., & Vilo, J. (2012). Robust rank aggregation for gene list integration and meta-analysis. Bioinformatics, 28(4), 573-580.
}
\author{
Yuanlong Hu
}
