% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_adjust.R
\name{score_adjust}
\alias{score_adjust}
\title{score_adjust}
\usage{
score_adjust(result, n = 100)
}
\arguments{
\item{result}{ScoreResult object}

\item{n}{integer.}
}
\description{
Adjust Score for Permutation Test
}
