% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_immpc.R
\name{score_immpc}
\alias{score_immpc}
\title{score_immpc}
\usage{
score_immpc(FP, disease_network, target)
}
\arguments{
\item{disease_network}{A data frame of disease network.This data frame containing a symbolic edge list in the first two columns.}

\item{target}{A data frame of drug target.}

\item{disease_biomarker}{A character.}

\item{geneset}{A list.}
}
\value{
ScoreResult object
}
\description{
Calculate the similarity of immune fingerprint and the characteristics of network topology
}
\author{
Yuanlong Hu
}
