% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_network.R
\name{score_network}
\alias{score_network}
\title{score_network}
\usage{
score_network(targetlist, disease_network, n = 100)
}
\arguments{
\item{targetlist}{A list containing drug target and disease biomarker.}

\item{disease_network}{A data frame of disease network containing two columns.}

\item{n}{The number of times random permutation sampling.}
}
\value{
ScoreResultNet object
}
\description{
Calculate the network score
}
\author{
Yuanlong Hu
}
