% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{ScoreFP-class}
\alias{ScoreFP-class}
\title{Class "ScoreFP"
This class represents the pathway fingerprint.}
\description{
Class "ScoreFP"
This class represents the pathway fingerprint.
}
\section{Slots}{

\describe{
\item{\code{Fingerprint}}{pathway fingerprint}

\item{\code{FPType}}{pathway fingerprint type}
}}

\author{
Yuanlong Hu
}
