% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{ScoreResult-class}
\alias{ScoreResult-class}
\title{Class "ScoreResult"
This class represents the score result.}
\description{
Class "ScoreResult"
This class represents the score result.
}
\author{
Yuanlong Hu
}
