% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFP.R
\name{simFP}
\alias{simFP}
\title{simFP}
\usage{
simFP(FP)
}
\arguments{
\item{FP}{A ScoreFP object}
}
\value{
a matrix
}
\description{
Calculate the similarity between Drug pathway Fingerprints
}
\examples{

  data("drugResult")
  sim_mat <- simFP(drugResult$demoFP)
}
\author{
Yuanlong Hu
}
