% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_network.R
\name{subset_network.BasicData}
\alias{subset_network.BasicData}
\title{subset_network}
\usage{
subset_network.BasicData(BasicData, from, to = NULL)
}
\arguments{
\item{BasicData}{A BasicData object.}

\item{from}{The source vertex.}

\item{to}{The target vertex.}
}
\value{
a BasicData object.
}
\description{
Subgraph of a Drug graph
}
\author{
Yuanlong Hu
}
