% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/subset_network.R
\name{subset_network}
\alias{subset_network}
\alias{subset_network,BasicData-method}
\title{subset_network}
\usage{
subset_network(BasicData, from, to = NULL)

\S4method{subset_network}{BasicData}(BasicData, from = "drug", to = NULL)
}
\arguments{
\item{BasicData}{A BasicData object.}

\item{from}{The source vertex.}

\item{to}{The target vertex.}
}
\value{
A BasicData object.
}
\description{
Subgraph of a Drug graph
}
\author{
Yuanlong Hu
}
