% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{tail,ScoreResult-method}
\alias{tail,ScoreResult-method}
\title{tail method}
\usage{
\S4method{tail}{ScoreResult}(x, ...)
}
\arguments{
\item{x}{a object of class ScoreResult}

\item{...}{other arguments}
}
\description{
tail method
}
\author{
Yuanlong Hu
}
