% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{tail}
\alias{tail}
\alias{tail,ScoreResult-method}
\title{Return the last Parts of a ScoreResult Object}
\usage{
\S4method{tail}{ScoreResult}(x, ...)
}
\arguments{
\item{x}{A ScoreResult object}

\item{...}{other arguments}
}
\description{
Return the last Parts of a ScoreResult Object
}
