% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_geneset.R
\name{to_biodescr}
\alias{to_biodescr}
\title{to_biodescr}
\usage{
to_biodescr(BioDescr)
}
\arguments{
\item{BioDescr}{A BioDescr object.}
}
\value{
A list.
}
\description{
Convert BioDescr object to a list of adjacency matrix
}
\examples{
\dontrun{
  to_biodescr(BioDescr)
}
}
\author{
Yuanlong Hu
}
