% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_geneset.R
\name{to_df}
\alias{to_df}
\title{to_df}
\usage{
to_df(list)
}
\arguments{
\item{list}{A list containing gene sets.}
}
\value{
A data frame.
}
\description{
Convert list to data.frame
}
\examples{
\dontrun{
  to_df(list)
}
}
\author{
Yuanlong Hu
}
