% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_geneset.R
\name{to_list}
\alias{to_list}
\title{to_list}
\usage{
to_list(dataframe, input = "single", sep = ", ")
}
\arguments{
\item{dataframe}{a data frame of 2 column with term/drug and gene}

\item{input}{one of the single or basket}

\item{sep}{When 'input' is 'basket'.}
}
\value{
list
}
\description{
Create a new list from a data.frame of drug target and disease biomarker as input
}
\examples{
\dontrun{
  to_list(dataframe)
}
}
\author{
Yuanlong Hu
}
