% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_network.R
\name{unionnet}
\alias{unionnet}
\title{unionnet}
\usage{
unionnet(...)
}
\arguments{
\item{...}{Drug graph from \code{PrepareData}.}
}
\value{
A new igraph object.
}
\description{
Union of graphs created by \code{PrepareData}
}
\author{
Yuanlong Hu
}
