% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_fp.R
\name{viewpathway}
\alias{viewpathway}
\title{viewpathway}
\usage{
viewpathway(x, drug, pathway_view = "intersect", gene_view = "drug")
}
\arguments{
\item{x}{a ScoreFP object.}

\item{drug}{character vector of drug name.}

\item{pathway_view}{Look at whether the pathway comes from one of intersect, drug or disease.}

\item{gene_view}{Look at whether the gene comes from one of intersect, drug or disease.}
}
\value{
ScoreResult object
}
\description{
View pathway information.
}
\author{
Yuanlong Hu
}
