% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_geneset.R
\name{write_gmt}
\alias{write_gmt}
\title{write_gmt}
\usage{
write_gmt(geneset, gmt_file)
}
\arguments{
\item{geneset}{A data.frame of 2 column with term/drug and gene.}

\item{gmt_file}{A character of gmt file name.}
}
\value{
gmt file
}
\description{
prints data frame to a gmt file
}
\author{
Yuanlong Hu
}
