% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BasicData-class}
\alias{BasicData-class}
\title{Class "BasicData"
This class represents the basic input data.}
\description{
Class "BasicData"
This class represents the basic input data.
}
\section{Slots}{

\describe{
\item{\code{BasicData}}{Alist containing basic data.}

\item{\code{Key}}{Column name of basic data.}

\item{\code{Relationship}}{Relationship.}
}}

\author{
Yuanlong Hu
}
