% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBasicData.R
\name{CreateBasicData}
\alias{CreateBasicData}
\title{CreateBasicData}
\usage{
CreateBasicData(...)
}
\arguments{
\item{...}{A data frame from \code{PrepareData}.}
}
\value{
a list
}
\description{
Prepare input data.
}
\examples{

  data("drugSample")
  drug_target <- PrepareData(drugSample$herb_target, col1 = "herb", col2 = "target")
  drug_target <- CreateBasicData(drug_target)
}
\author{
Yuanlong Hu
}
