% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBasicData.R
\name{CreateBasicData}
\alias{CreateBasicData}
\title{CreateBasicData}
\usage{
CreateBasicData(...)
}
\arguments{
\item{...}{A data frame from \code{PrepareData}.}
}
\value{
a list
}
\description{
Prepare input data.
}
\examples{

  data("drugSample")
  drug_herb <- PrepareData(drugSample$drug_herb, col1 = "drug", col2 = "herb")
  herb_target <- PrepareData(drugSample$herb_target,
                             col1 = "herb", col2 = "target",
                             format = "basket", sep = ", ")
  drug_target <- CreateBasicData(drug_herb, herb_target)
}
\author{
Yuanlong Hu
}
