% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBasicData.R
\name{CreateBasicData}
\alias{CreateBasicData}
\title{CreateBasicData}
\usage{
CreateBasicData(drug_herb, herb_target)
}
\arguments{
\item{drug_herb}{A data frame or listcontaining prescription and herb}

\item{herb_target}{A data.frame or list containing herb and target}
}
\value{
a list
}
\description{
Prepare input data.
}
\author{
Yuanlong Hu
}
