% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDisDrugNet.R
\name{CreateDisDrugNet}
\alias{CreateDisDrugNet}
\title{CreateDisDrugNet}
\usage{
CreateDisDrugNet(BasicData, drug, disease)
}
\arguments{
\item{BasicData}{BasicData object.}

\item{drug}{Character vector, the drug。}

\item{disease}{Character vector, the disease.}
}
\value{
A BasicData object.
}
\description{
Create Disease-Drug Network
}
\author{
Yuanlong Hu
}
