% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDisDrugNet.R
\name{exportCytoscape}
\alias{exportCytoscape}
\title{exportCytoscape}
\usage{
exportCytoscape(BasicData, file)
}
\arguments{
\item{BasicData}{BasicData object.}

\item{file}{file}
}
\value{
A workbook object
}
\description{
Export an xlsx file to Cytoscape.
}
\author{
Yuanlong Hu
}
