% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrFP.R
\name{extrFP}
\alias{extrFP}
\title{extrFP}
\usage{
extrFP(disease_biomarker, drug_target, geneset = "ImmGenTop150")
}
\arguments{
\item{disease_biomarker}{A character.}

\item{drug_target}{A data frame of drug target.}

\item{geneset}{one of "ImmGenTop150" and "KEGG"}
}
\value{
ScoreFP object
}
\description{
Calculate the Drug Fingerprints
}
\author{
Yuanlong Hu
}
