% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{HerbResult-class}
\alias{HerbResult-class}
\title{Class \code{HerbResult}
This class represents the biological descriptor data.}
\description{
Class \code{HerbResult}
This class represents the biological descriptor data.
}
\section{Slots}{

\describe{
\item{\code{Drug_Herb}}{Data frame, Drug-herb relationship.}

\item{\code{Herb_Herb}}{Herb-herb association Rule Graph, it is a directed graph.}
}}

\author{
Yuanlong Hu
}
