% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immcp-package.R
\docType{package}
\name{immcp-package}
\alias{immcp-package}
\alias{immcp}
\title{screen and analysis and candidate prescription related to disease}
\description{
The package implements methods to analyze and screen candidate prescription
related to disease based pathway fingerprints.
}
