% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_characterstics.R
\name{natural_connectivity}
\alias{natural_connectivity}
\title{natural_connectivity}
\usage{
natural_connectivity(graph)
}
\arguments{
\item{graph}{The graph.}
}
\value{
a numeric vector
}
\description{
Calculate the natural connectivity
}
\examples{

data("drugSample")
graph <- igraph::graph.data.frame(drugSample$disease_network)
natural_connectivity(graph)
}
\author{
Yuanlong Hu
}
