% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_characterstics.R
\name{network_char}
\alias{network_char}
\title{network_char}
\usage{
network_char(graph, total_network = FALSE)
}
\arguments{
\item{graph}{The graph.}

\item{total_network}{Calculate for total network or each nodes.}
}
\value{
a data frame
}
\description{
Calculate the network characters
}
\examples{

data("drugSample")
graph <- igraph::graph.data.frame(drugSample$disease_network)
network_char(graph)
}
\author{
Yuanlong Hu
}
