% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_node_ks.R
\name{network_node_ks}
\alias{network_node_ks}
\title{network_node_ks}
\usage{
network_node_ks(graph1, graph2, replicate = 1000)
}
\arguments{
\item{graph1}{A igraph object.}

\item{graph2}{A igraph object.}

\item{replicate}{Number vector, the number of conduct bootstrapping sampling replications.}
}
\value{
A data frame
}
\description{
Kolmogorov-Smirnov tests for node characters between networks
}
\author{
Yuanlong Hu
}
