% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_characterstics.R
\name{network_node_ks}
\alias{network_node_ks}
\title{network_node_ks}
\usage{
network_node_ks(network_char, replicate = 1000)
}
\arguments{
\item{network_char}{list; network node characters}

\item{replicate}{the number of conduct bootstrapping sampling replications}
}
\value{
a data frame
}
\description{
Kolmogorov-Smirnov tests for node characters between networks
}
\examples{

data("drugSample")
graph <- igraph::graph.data.frame(drugSample$disease_network)
network_char <- network_char(graph)
network_node_ks(network_char)
}
\author{
Yuanlong Hu
}
