% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDisDrugNet.R
\name{plot_BasicData}
\alias{plot_BasicData}
\title{plot_BasicData}
\usage{
plot_BasicData(
  BasicData,
  drug = NULL,
  disease = NULL,
  color = c(drug = "blue", herb = "lightblue", target = "orange"),
  ...
)
}
\arguments{
\item{BasicData}{BasicData object.}

\item{drug}{drug}

\item{disease}{disease}

\item{color}{Color}

\item{...}{Arguments}
}
\value{
Returns NULL, invisibly.
}
\description{
Plot Disease-Drug Network
}
\examples{
data(drugdemo)
drug_herb <- PrepareData(drugdemo$drug_herb, from = "drug", to="herb")
herb_compound <- PrepareData(drugdemo$herb_compound, from = "herb", to="compound")
compound_target <- PrepareData(drugdemo$compound_target, from = "compound", to="target")
disease <- PrepareData(drugdemo$disease, diseaseID = "disease",from = "target", to="target")
BasicData <- CreateBasicData(drug_herb, herb_compound, compound_target, diseasenet = disease)
DisDrugNet <- CreateDisDrugNet(BasicData, drug = "Drug1", disease = "disease")
plot_BasicData(DisDrugNet)
}
\author{
Yuanlong Hu
}
