% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{plot_density}
\usage{
plot_density(result, drug, view = "Change_average_degree", fill = "#6495ED")
}
\arguments{
\item{result}{an object of class ScoreResult.}

\item{drug}{a character of drug name.}

\item{view}{the index}

\item{fill}{fill color.}
}
\value{
a ggplot
}
\description{
Plot smoothed density estimates for adjusted score
}
\author{
Yuanlong Hu
}
