% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/CreateDisDrugNet.R
\name{plot_graph}
\alias{plot_graph}
\alias{plot_graph,BasicData-method}
\alias{plot_graph,igraph-method}
\title{Plot Disease-Drug Network}
\usage{
plot_graph(
  graph,
  drug,
  disease,
  vis = "visNetwork",
  color = c(drug = "blue", herb = "lightblue", target = "orange"),
  width = 1,
  ...
)

\S4method{plot_graph}{BasicData}(
  graph,
  drug,
  disease,
  vis = "visNetwork",
  color = c(drug = "blue", herb = "lightblue", target = "orange"),
  width = 1,
  ...
)

\S4method{plot_graph}{igraph}(
  graph,
  drug,
  disease,
  vis = "visNetwork",
  color = c(drug = "blue", herb = "lightblue", target = "orange"),
  width = 1,
  ...
)
}
\arguments{
\item{graph}{graph.}

\item{drug}{drug.}

\item{disease}{disease.}

\item{vis}{one of "igraph", "visNetwork" and "shiny".}

\item{color}{Color}

\item{width}{width}

\item{...}{Arguments}
}
\value{
Returns NULL, invisibly.
}
\description{
Plot Disease-Drug Network
}
\author{
Yuanlong Hu
}
