% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{plot_network}
\usage{
plot_network(ResNet, Drug, visIgraphLayout = "none")
}
\arguments{
\item{ResNet}{A ScoreResultNet object}

\item{Drug}{A character of drug name}

\item{visIgraphLayout}{use layout of igraph}
}
\value{
visNetwork
}
\description{
Drug-Target network visualization
}
\author{
Yuanlong Hu
}
