% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBasicData.R
\name{PrepareData}
\alias{PrepareData}
\title{PrepareData}
\usage{
PrepareData(data, col1, col2, format = "single", sep)
}
\arguments{
\item{data}{A data frame containing interaction information.}

\item{col1}{A charactor containing "drug", "herb", "compound", or "target".}

\item{col2}{A character containing "drug", "herb", "compound", or "target".}

\item{format}{one of "single" or "basket".}

\item{sep}{Separator.}
}
\value{
a list
}
\description{
Prepare input format.
}
\examples{

  data("drugSample")
  drug_target <- PrepareData(drugSample$herb_target, col1 = "herb", col2 = "target")
}
\author{
Yuanlong Hu
}
