% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_geneset.R
\name{read_gmt}
\alias{read_gmt}
\title{write_gmt}
\usage{
read_gmt(gmtfile, out_type = "data.frame")
}
\arguments{
\item{gmtfile}{A GMT file name or URL containing gene sets.}

\item{out_type}{A character vector of object name. one of "data.frame", "list", "GeneSetCollection"}
}
\value{
data.frame, list or GeneSetCollection
}
\description{
parse gmt file to a data.frame
}
\author{
Yuanlong Hu
}
