% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list.R
\name{read_gmt}
\alias{read_gmt}
\title{write_gmt}
\usage{
read_gmt(gmtfile, input_list = FALSE)
}
\arguments{
\item{gmtfile}{gmt file}
}
\value{
data.frame
}
\description{
parse gmt file to a data.frame
}
\author{
Yuanlong Hu
}
