% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_fp.R
\name{score_fp}
\alias{score_fp}
\title{score_fp}
\usage{
score_fp(FP, n = 100, two_tailed = TRUE)
}
\arguments{
\item{FP}{A ScoreFP.}

\item{n}{The number of permutations.}

\item{two_tailed}{a logical: select a two-tailed p-value}
}
\value{
ScoreResult
}
\description{
Calculate the pathway fingerprint.similarity between disease and prescription.
}
\examples{

  data("drugResult")
  res <- score_fp(drugResult$demoFP, n=100)
}
\references{
Ye, H., Tang, K., Yang, L., Cao, Z., & Li, Y. (2012). Study of drug function based on similarity of pathway fingerprint. Protein & cell, 3(2), 132-139.
}
\author{
Yuanlong Hu
}
