% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_fp.R
\name{score_fp}
\alias{score_fp}
\title{score_immpc}
\usage{
score_fp(FP, n = 100)
}
\arguments{
\item{FP}{A ScoreFP.}

\item{n}{The number of times random permutation sampling.}
}
\value{
ScoreResult object
}
\description{
Calculate the pathway fingerprint.similarity between disease and prescription.
}
\examples{
\dontrun{
  data("drugSample")
  FP <- extrFP(disease_biomarker = drugSample$disease_biomarker,
               drug_target = drugSample$herb_target,
               geneset = "ImmGenTop150")
  res <- score_fp(FP, n=100)
  res <- as.data.frame(res)
}
}
\author{
Yuanlong Hu
}
