% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_immpc.R
\name{score_immpc}
\alias{score_immpc}
\title{score_immpc}
\usage{
score_immpc(disease_biomarker, disease_network, target, geneset = NULL)
}
\arguments{
\item{disease_network}{A data frame of disease network.This data frame containing a symbolic edge list in the first two columns.}

\item{target}{A data frame of drug target.}
}
\value{
ScoreResult object
}
\description{
Calculate the similarity of immune fingerprint and the characteristics of network topology
}
\author{
Yuanlong Hu
}
