% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_network.R
\name{score_network}
\alias{score_network}
\title{score_network}
\usage{
score_network(BasicData, n = 100)
}
\arguments{
\item{BasicData}{A BasicData object.}

\item{n}{Number vector, the number of times random permutation sampling.}
}
\value{
A list.
}
\description{
Calculate the network score
}
\author{
Yuanlong Hu
}
