% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_network.R
\name{score_network}
\alias{score_network}
\title{score_network}
\usage{
score_network(Tar, DNet, n = 100)
}
\arguments{
\item{Tar}{A list containing drug target and disease biomarker.}

\item{DNet}{A data frame of disease network containing two columns.}

\item{n}{The number of times random permutation sampling.}
}
\value{
ScoreResultNet object
}
\description{
Calculate the network score
}
\examples{
\dontrun{
  data("drugSample")
  Tar <- drugSample$herb_target
  disease_network <- drugSample$disease_network
  res <- score_network(Tar, DNet)
  res <- as.data.frame(res)
}
}
\author{
Yuanlong Hu
}
