% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_rule.R
\name{score_rule}
\alias{score_rule}
\title{score_rule}
\usage{
score_rule(BasicData, support = 0.1, confidence = 0.8, maxlen = 2)
}
\arguments{
\item{BasicData}{BasicData object.}

\item{support}{cutoff of support}

\item{confidence}{cutoff of confidence}

\item{maxlen}{maxlen}
}
\value{
Returns NULL, invisibly.
}
\description{
Plot Disease-Drug Network
}
\author{
Yuanlong Hu
}
