% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_sim.R
\name{score_sim}
\alias{score_sim}
\title{score_sim}
\usage{
score_sim(BioDescr, method = "jaccard", n = 1000)
}
\arguments{
\item{BioDescr}{BioDescr object.}

\item{method}{method, default "jaccard".}

\item{n}{number}
}
\value{
a list
}
\description{
Calculating drug-disease similarity based on biological descriptors
}
\examples{
data(drugdemo)
drug_herb <- PrepareData(drugdemo$drug_herb, from = "drug", to="herb")
herb_compound <- PrepareData(drugdemo$herb_compound, from = "herb", to="compound")
compound_target <- PrepareData(drugdemo$compound_target, from = "compound", to="target")
disease <- PrepareData(drugdemo$disease, diseaseID = "disease",from = "target", to="target")
BasicData <- CreateBasicData(drug_herb, herb_compound, compound_target, diseasenet = disease)
biodescr <- extr_biodescr(BasicData, geneset= "kegg")
res <- score_sim(biodescr, method="jaccard", n=100)
}
\author{
Yuanlong Hu
}
