% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_sim.R
\name{score_sim}
\alias{score_sim}
\title{score_sim}
\usage{
score_sim(BioDescr, method = "jaccard", n = 1000)
}
\arguments{
\item{BioDescr}{BioDescr object.}

\item{method}{method, default "jaccard".}

\item{n}{number}
}
\value{
a list
}
\description{
Simi
}
\author{
Yuanlong Hu
}
