% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{ScoreResult-class}
\alias{ScoreResult-class}
\title{Class "ScoreResult"}
\description{
This class represents the result of score.
}
\section{Slots}{

\describe{
\item{\code{ScoreResult}}{all score reslut.}

\item{\code{adj}}{distribution data}
}}

\author{
Yuanlong Hu
}
