% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFP.R
\name{simFP}
\alias{simFP}
\title{simFP}
\usage{
simFP(FP)
}
\arguments{
\item{FP}{A ScoreFP object}
}
\value{
a matrix
}
\description{
Calculate the similarity between Drug pathway Fingerprints
}
\examples{
\dontrun{
  data("drugSample")
  FP <- extrFP(disease_biomarker = drugSample$disease_biomarker,
               drug_target = drugSample$herb_target,
               geneset = "ImmGenTop150")
  sim_mat <- simFP(FP)
}
}
\author{
Yuanlong Hu
}
