% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_geneset.R
\name{to_list}
\alias{to_list}
\title{to_list}
\usage{
to_list(dataframe)
}
\arguments{
\item{dataframe}{a data frame of 2 column with term/drug and gene}
}
\value{
list
}
\description{
Create a new list from a data.frame of drug target and disease biomarker as input
}
\author{
Yuanlong Hu
}
