% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_fp.R
\name{viewpathway}
\alias{viewpathway}
\title{viewpathway}
\usage{
viewpathway(x, drug, view = "intersect")
}
\arguments{
\item{x}{a ScoreFP object.}

\item{drug}{character vector of drug name.}

\item{view}{one of intersect, drug or disease.}
}
\value{
ScoreResult object
}
\description{
View pathway information.
}
\author{
Yuanlong Hu
}
